Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main (){
-    int x;
-    printf("%d",&x);
+    float x;
+    printf("%f",&x);
 
     float area;
     area = 3.14*(r*r);
